+++ src/Cooked/MockChain.hs
@@ -14,7 +14,9 @@ import Cooked.MockChain.Staged as X hiding
   )
 import Cooked.MockChain.Testing as X
 import Cooked.MockChain.UtxoSearch as X
+import Cooked.MockChain.UtxoState as X (UtxoState)
 import Cooked.MockChain.UtxoState as X hiding
   ( UtxoPayload,
+    UtxoPayloadSet,
     UtxoState (UtxoState, utxoState),
   )